avatar-meeting-bot/requirements.txt
# Speech to text
openai-whisper

# Voice cloning
f5-tts

# Avatar — SadTalker dependencies
# (install SadTalker separately: git clone https://github.com/OpenTalker/SadTalker.git)
# then: pip install -r C:\Users\Administrator\SadTalker\requirements.txt
face_alignment==1.3.5
gfpgan
imageio
imageio-ffmpeg
librosa==0.9.2
resampy==0.3.1
scipy
kornia
yacs

# Audio capture
sounddevice
soundfile

# HTTP client (LLM)
requests

# Video display
opencv-python
